#318: Create an Ommph setup and fix a few minor project issues
This changes adds an Oomph based setup which follows the
"Getting Started" tutorial. It will allow the user to choose between
a minimal set as described in the tutorial or a full Kura developer
setup.

For this it was necessary to also fix a few project issues in order to
create a, mostly, error free workspace.
 
The steps to be taken for the end user are:

 * Install Java 8
 * Download and run the Eclipse installer
  * Select Eclipse IDE edition
  * Select "Eclipse Kura"
  * Enter location for installation
  * Wait a few minutes

In addition this change fixes a few project settings issues:

 * org.eclipse.soda.dk.comm package path (actually is
src/main/java/org/eclipse/soda/dk instead of
src/main/java/org/eclipse/soda/dk/comm)
 * Don't ignore Tycho and WTP maven plugins but install m2eclipse
adapters instead

Signed-off-by: Jens Reimann <jreimann@redhat.com>